+++ webpack.config.js
@@ -1,0 +1,56 @@
+const {resolve} = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+  // Bundle the transpiled code in dist
+  entry: {
+    library: resolve('./dist/index.js')
+  },
+
+  // Generate a bundle in dist folder
+  output: {
+    path: resolve('./dist'),
+    filename: 'bundle.js',
+    library: 'react-map-gl',
+    libraryTarget: 'umd'
+  },
+
+  // Exclude any non-relative imports from resulting bundle
+  externals: [
+    /^[a-z\-0-9]+$/
+  ],
+
+  // devtool: 'source-maps'
+
+  plugins: [
+    new webpack.optimize.UglifyJsPlugin({
+      output: {
+        comments: false
+      }
+    })
+  ]
+  /*
+  plugins: [
+    new webpack.optimize.AggressiveMergingPlugin(),
+    new webpack.optimize.UglifyJsPlugin({
+      sourceMap: false,
+      compress: {
+        sequences: true,
+        dead_code: true,
+        conditionals: true,
+        booleans: true,
+        unused: true,
+        if_return: true,
+        join_vars: true,
+        drop_console: true
+      },
+      mangle: {
+        except: ['$super', '$', 'exports', 'require']
+      },
+      output: {
+        comments: false
+      }
+    })
+  ]
+  */
+};